desloppify: move fixed volume objective module
